requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
passlib[bcrypt]
email-validator